+++ plugins/qmljs/3rdparty/qtcreator-libs/qmljs/qmljsdialect.h
@@ -72,7 +72,7 @@ private:
     Enum m_dialect;
 };
 
-QMLJS_EXPORT uint qHash(const Dialect &o);
+QMLJS_EXPORT size_t qHash(const Dialect &o);
 
 QMLJS_EXPORT QDebug operator << (QDebug &dbg, const Dialect &dialect);
